split out types
