+++ x/ccv/types/proposal_test.go
@@ -16,7 +16,7 @@ import (
 
 	clienttypes "github.com/cosmos/ibc-go/v4/modules/core/02-client/types"
 	ibctmtypes "github.com/cosmos/ibc-go/v4/modules/light-clients/07-tendermint/types"
-	"github.com/cosmos/interchain-security/x/ccv/types"
+	"github.com/cosmos/interchain-security/x/types"
 )
 
 func TestConsumerAdditionProposalValidateBasic(t *testing.T) {
